Implement Order creation feature with MVC architecture (Controller, Repository, Service) and add getAllOrders method
