--- src/main/java/com/adobe/bookstore/BookStock.java
@@ -1,48 +1,0 @@
-package com.adobe.bookstore;
-
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "book_stock")
-@JsonSerialize
-public class BookStock {
-
-    @Id
-    @Column(name = "id", nullable = false)
-    private String id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "quantity", nullable = false)
-    private Integer quantity;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getQuantity() {
-        return quantity;
-    }
-
-    public void setQuantity(Integer quantity) {
-        this.quantity = quantity;
-    }
-}
