+++ src/main/java/com/adobe/bookstore/BookstoreApplication.java
@@ -2,12 +2,14 @@ package com.adobe.bookstore;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan("com.adobe")
 public class BookstoreApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(BookstoreApplication.class, args);
-    }
+  public static void main(String[] args) {
+    SpringApplication.run(BookstoreApplication.class, args);
+  }
 
 }
