+++ src/main/java/com/adobe/bookstore/model/order/constants/Status.java
@@ -1,0 +1,7 @@
+package com.adobe.bookstore.model.order.constants;
+
+public enum Status {
+  Success,  // Order placed
+  Pending,  // Checking stock availability
+  Rejected  // No stock available
+}
