+++ src/main/java/com/adobe/bookstore/repository/BookStockRepository.java
@@ -1,8 +1,10 @@
-package com.adobe.bookstore;
+package com.adobe.bookstore.repository;
 
+import com.adobe.bookstore.model.BookStock;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface BookStockRepository extends JpaRepository<BookStock, String> {
+
 }
