--- src/main/java/com/adobe/error/GlobalExceptionHandler.java
@@ -1,15 +1,0 @@
-package com.adobe.error;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-  @ExceptionHandler(OrderNotFoundException.class)
-  public ResponseEntity<String> handleOrderNotFoundException(OrderNotFoundException e) {
-    return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-  }
-}
